backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic>=2.10.4
pydantic-settings>=2.3.0
groq>=0.9.0
tavily-python==0.3.3
firecrawl-py==0.0.16
browser-use==0.1.40
playwright>=1.49.0
PyGithub==2.3.0
langchain-groq>=0.1.3
langchain-anthropic==0.3.3
sse-starlette==2.1.0
httpx>=0.27.2
fpdf2==2.7.9
python-dotenv==1.0.1
tiktoken>=0.7.0
aiofiles==23.2.1
python-multipart==0.0.9
